Solver fixes
  - P1dB refinement degraded instead of raising, which was discarding a
    complete sweep's artifacts when one mid-bracket power would not
    converge; it also now receives the caller's continuation settings.
  - The signal-power continuation fallback no longer sizes its fixed-step
    count as span/min_step, which had scheduled 6881 steps (~29 h) for one
    power point.
  - build_effective_snail_line stamped the SNAIL small-signal stiffness
    into K while the branch law also supplied it through Bphi @ i_branch.
    The double count halved the line inductance and stopped the benchmark
    amplifying at all. Scoped to that builder; ipm and jc_doc are correct
    and 2c is unaffected.

Circuit provenance
  - outputs/ipm_python_design is not the fabricated 2c device. 20 scripts
    now point at designs/ipm_2c_fixed and 358 stale-circuit outputs are
    removed.

2c against measurement
  - Operating point located at 7.100 GHz / 7.2311 uA, rms 1.247 dB.
  - Gain band matches: 4.375 vs 4.268 GHz wide, 7.062 vs 7.080 GHz centre,
    0.92 dB envelope deficit. Ripple of 3-4 dB peak-to-peak makes any
    few-frequency comparison unreliable.
  - Compression at seven frequencies; the model saturates late by 2.86 dB
    on average, which is the direction fabrication defects predict.

Le Gal benchmark
  - Assembled dispersion validated to 1.1e-15 against the builder relation.
  - HB Kerr confirmed consistent with its own branch law to 4.4%.
  - After the stiffness fix the line amplifies: 13.85 dB at 6.4 GHz and
    13.82 at 8.6, against 0.12 and 0.08 before, sideband converged.
  - Compression is accompanied by real pump depletion, and P1dB varies far
    more than pure depletion permits.

Presentation materials for 2c and JTWPA are in outputs/presentation with
docs/development/presentation_numbers_2c_jtwpa.md.

The pump-power gap is deliberately not asserted: the Norton termination
makes the available power I^2*Z/8 rather than the I^2*Z/2 used by the
conversion, a 6.02 dB difference that is not yet resolved.

+++ src/twpa_solver/multitone/io.py
@@ -45,8 +45,11 @@ def write_compression_outputs(
             value = np.asarray(state)
             np.savez_compressed(
                 destination / f"multitone_solution_{name}.npz",
-                X_real=np.asarray(value.real, dtype=np.float32),
-                X_imag=np.asarray(value.imag, dtype=np.float32),
+                # Validation and continuation diagnostics need the converged
+                # state at the solver's precision; float32 quantizes tiny
+                # signal sectors enough to spoil JVP finite differences.
+                X_real=np.asarray(value.real, dtype=np.float64),
+                X_imag=np.asarray(value.imag, dtype=np.float64),
             )
     if spatial_rows is not None:
         columns = sorted({key for row in spatial_rows for key in row})
